refactor(keyboard-control): split gui helpers from node
